requirements.txt
# Django and DRF
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5

# Database
psycopg2-binary>=2.9.9

# Cache
redis>=5.0.0
django-redis>=5.4.0

# Server & Static
gunicorn>=21.2.0
whitenoise>=6.6.0

# Environment
python-dotenv>=1.0.0

# Utilities
Pillow>=10.2.0
python-dateutil>=2.8.2
requests>=2.31.0

# Docker
# docker>=7.0.0
